+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Enhancements
 
+- [Backup NG] Restore logs moved to restore tab [#3772](https://github.com/vatesfr/xen-orchestra/issues/3772) (PR [#3802](https://github.com/vatesfr/xen-orchestra/pull/3802))
+
 ### Bug fixes
 
 ### Released packages